New branch with UUID hf78qniafdtfijhripzwnfhsq
Initial checkout location: /home/mnt/distfiles/maint

Initial version

Distfiles redirection framework: The purpose of this framework is to
avoid duplication of identical source package information metadata
files, source file archives and patch files across multiple filesystems.

The idea is to move all the real files to THIS filesystem and replace
the original files with symlinks to the new locations.

As a result, multiple copies of the same source archive will be replaced
by a single copy within THIS filesystem, avoiding duplication.

For ease of use, the created replacement symlinks will contain the
absolute path to THIS filesystem's mount point.

While this allows to move the directories containing the replacement
symlinks to new locations without invalidating the symlinks, it also
means that the mount point most not be changed or all the replacement
symlinks will need adjustment before they can be used again.